読み込み完成 必要なリソースを追加

+++ src/sysFrame/DirectX/Heap/Heap.cpp
@@ -20,12 +20,12 @@ bool Heap::Create(ID3D12Device* device, D3D12_DESCRIPTOR_HEAP_TYPE type, uint32_
 
 D3D12_CPU_DESCRIPTOR_HANDLE Heap::GetCPUHandle(uint32_t index) const {
     D3D12_CPU_DESCRIPTOR_HANDLE handle = heap_->GetCPUDescriptorHandleForHeapStart();
-    handle.ptr += device_->GetDescriptorHandleIncrementSize(type_) * index;
+    handle.ptr += static_cast<SIZE_T>(device_->GetDescriptorHandleIncrementSize(type_)) * index;
     return handle;
 }
 
 D3D12_GPU_DESCRIPTOR_HANDLE Heap::GetGPUHandle(uint32_t index) const {
     D3D12_GPU_DESCRIPTOR_HANDLE handle = heap_->GetGPUDescriptorHandleForHeapStart();
-    handle.ptr += device_->GetDescriptorHandleIncrementSize(type_) * index;
+    handle.ptr += static_cast<SIZE_T>(device_->GetDescriptorHandleIncrementSize(type_)) * index;
     return handle;
 }
